% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyberpunk_geoms.R
\name{geom_glowing_area}
\alias{geom_glowing_area}
\title{A Cyberpunk styled area chart}
\usage{
geom_glowing_area(alpha =  1, size = 1.5, glow_alpha = 0.05, layers = 10, glow_size = .65)
}
\arguments{
\item{alpha}{set the alpha level for the base line.}

\item{size}{set the base line size.}

\item{glow_alpha}{alpha level for the glow effect}

\item{layers}{number of glowing layers to draw,}

\item{glow_size}{size of the glow}
}
\value{
A list of ggplot geoms containin geom_area, geom_lines, and geom_point.
}
\description{
Glowing Lines, lightly colored area plots for a cyberpunk area plot.
}
\details{
Create a cyberpunk themed chart using ggplot. This plot is a fancy area plot, that uses a glow effect around the upper line.
}
\examples{
\dontrun{
library(ggplot2)
library(tidyverse)

df = data.frame(A=c(1,4,4,6,7,5,1),
               B=c(4,3,5,7,5,6,7),
               I=c(1,2,3,4,5,6,7)) %>%
pivot_longer(cols = c(A,B),names_to = "group", values_to = "count")


df %>%
   ggplot(aes(x=I, y = count, color = group, fill= group))+
     geom_glowing_area()+
     theme_cyberpunk()+
     scale_fill_manual(values = c("#08F7FE", "#FE53BB"))+
     scale_color_manual(values = c("#08F7FE", "#FE53BB"))
}
}
