% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyberpunk_geoms.R
\name{geom_glowing_line}
\alias{geom_glowing_line}
\alias{geom_linesaber}
\title{A lightsaber inspired plot}
\usage{
geom_glowing_line(
  alpha = 1,
  size = 1.5,
  glow_alpha = 0.05,
  layers = 10,
  glow_size = 0.65
)

geom_linesaber(
  alpha = 1,
  size = 1.25,
  glow_alpha = 0.05,
  layers = 10,
  glow_size = 0.65,
  inner_color = "white"
)
}
\description{
A lightsaber inspired plot
}
