% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Col and pal.R
\name{get_cyberpunk_colors}
\alias{get_cyberpunk_colors}
\title{Extract a list of color codes}
\usage{
get_cyberpunk_colors(c(color1, color2))
}
\value{
A list of color codes
}
\description{
Mostly an internal function, it serves to extract colors from \link[ggCyberPunk]{colors_cyberpunk}
}
