% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_and_pal.R
\name{get_cyberpunk_palettes}
\alias{get_cyberpunk_palettes}
\title{get cyberpunk palette from list.}
\usage{
get_cyberpunk_palettes(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{name of the palette}

\item{reverse}{should the palette be reversed}

\item{...}{arguments}
}
\description{
get cyberpunk palette from list.
}
