% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Col and pal.R
\docType{data}
\name{palettes_cyberpunk}
\alias{palettes_cyberpunk}
\title{List of Palettes in ggCyberPunk}
\format{An object of class \code{list} of length 4.}
\usage{
palettes_cyberpunk[pal_name]
}
\arguments{
\item{pal_name}{name of the palette to use}
}
\value{
a list of names
}
\description{
A collection of palettes used in ggCyberPunk.
}
\details{
'main' is the default palette used. It has the neon blue, pink, yellow, and orange
'zune' is a two color palette using the old zune orange and pink
'cotton candy' is a blue and pink two color palette
'laser sword' contains colors best suited for \link[ggCyberPunk]{geom_linesaber} and is based off light saber colors.
}
\keyword{datasets}
