% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Col and pal.R
\name{scale_color_cyberpunk}
\alias{scale_color_cyberpunk}
\title{Color with CyberPunk palettes!}
\usage{
scale_color_cyberpunk(palette = "main", discrete = T, reverse = F, ...)
}
\description{
Color with CyberPunk palettes!
}
