% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyberpunk_themes.R
\name{theme_cyberpunk}
\alias{theme_cyberpunk}
\title{Cyberpunk Themes for Sci-Fi Obsessed Teans}
\usage{
theme_cyberpunk( font = "Source Code Pro", main.text.color = "#EE9537",
sub.text.color = "#EE9537", base.size = 15, plot.background.color = "#212946",
legend.position = "right", panel.border.color = "#494F67", panel.background = "#212946",
panel.grid.color = "#EE9537", axis.text.color = "#EE9537", axis.text.size = base.size * 0.6666667,
subtitle.size = base.size * 0.8, legend.text = base.size * 0.8, legend.title = base.size * 0.85,
axis.title.size = base.size * 0.8, title.size = 15)
}
\arguments{
\item{font}{sets the default font}

\item{main.text.color}{sets the main text color}

\item{sub.text.color}{sets the subtext color}

\item{base.size}{sets the base font size}

\item{plot.background.color}{sets the background color}

\item{legend.position}{sets the default legend position}

\item{panel.border.color}{sets the panel border color}

\item{panel.background}{sets the background color for the plot}

\item{panel.grid.color}{sets the grid color for the plot}

\item{axis.text.color}{sets the axis text color}

\item{axis.text.size}{sets the axis text size}

\item{subtitle.size}{sets the  subtitle size}

\item{legend.text}{sets the title text size}

\item{legend.title}{sets the legend title size}

\item{axis.title.size}{sets the axis title size,}

\item{title.size}{sets the title size}
}
\description{
A theme to make beautiful cyer punk themed plots. Based of the theme newRetro in the vapoRwave package
}
\details{
A theme designed to work with the glowing lines plots to provide a cyber punk feel.
}
