% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Col and pal.R
\docType{data}
\name{colors_cyberpunk}
\alias{colors_cyberpunk}
\title{Named colors for ggCyberPunk}
\format{An object of class \code{character} of length 11.}
\usage{
colors_cyberpunk[name]
}
\arguments{
\item{name}{name of the color used}
}
\value{
A named character containing a hex color code.
}
\description{
A set of named colors for use in this package's palettes.
}
\keyword{datasets}
