% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyberpunk geoms.R
\name{geom_glowing_line}
\alias{geom_glowing_line}
\title{Sci-Fi themed line plots}
\usage{
geom_glowing_line <- function(alpha =  1, size = 1)
}
\arguments{
\item{alpha}{set the alpha level for the base line.}

\item{size}{set the base line size.}

\item{glow_alpha}{set the alpha for the glow}
}
\value{
A list of ggplot geoms containing, geom_lines, and geom_point. geom_linesaber does not return a geom_point
}
\description{
Glowing Lines, lightly colored area plots for a cyberpunk/Laser sword themed line plot
}
\details{
Create a cyberpunk themed chart using ggplot. This plot is a fancy line plot, that uses a glow effect around the line.
}
\examples{
library(ggplot2)
library(tidyverse)

df = data.frame(A=c(1,4,4,6,7,5,1),
               B=c(4,3,5,7,5,6,7),
               I=c(1,2,3,4,5,6,7)) \%>\%
pivot_longer(cols = c(A,B),names_to = "group", values_to = "count")


df \%>\%
   ggplot(aes(x=I, y = count, color = group, fill= group))+
     geom_glowing_area()+
     theme_dark()+
     scale_fill_manual(values = c("#08F7FE", "#FE53BB"))+
     scale_color_manual(values = c("#08F7FE", "#FE53BB"))
}
