% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Col and pal.R
\name{get_cyberpunk_palettes}
\alias{get_cyberpunk_palettes}
\title{get cyberpunk palletes from list.}
\usage{
get_cyberpunk_palettes(palette = "main", reverse = FALSE, ...)
}
\description{
get cyberpunk palletes from list.
}
